pyproject.toml
[project]
name = "unsupervised-persona-elicitation"
version = "0.1.0"
description = "Unsupervised Elicitation of Language Models"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "adjustText",
    "alembic",
    "altair",
    "anthropic",
    "cattrs",
    "datasets",
    "fastapi==0.100.0",
    "fire",
    "hydra-core",
    "matplotlib",
    "mistralai",
    "openai==0.28.0",
    "pandas",
    "pebble",
    "polars",
    "pre-commit",
    "pydantic",
    "python-dotenv>=1.2.1",
    "replicate",
    "requests>=2.32.5",
    "scikit-learn",
    "scipy",
    "seaborn",
    "sqlalchemy==2.0.18",
    "tabulate",
    "tenacity",
    "termcolor",
    "tiktoken",
    "tqdm",
    "trueskill",
    "typer",
    "uvicorn[standard]==0.22.0",
    "wandb",
]

[project.scripts]
icm = "src.experiments.ICM:entry_point"

[tool.hatch.build.targets.wheel]
packages = ["src", "core"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 80
indent-width = 4

# Assume Python 3.12
target-version = "py310"

# [tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.

# select = ["E4", "E7", "E9", "F"]
# ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
# fixable = ["ALL"]
# unfixable = []

# Allow unused variables when underscore-prefixed.
# dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.pydocstyle]
convention = "google"
